Remove stray text
I think the text I removed must have been added by mistake? It is out of place, and for some reason accessing the latest version of this page through google translate on a chromium browser results in a load of mojibake. All the other pages are fine, so this is my only theory on why that would be.